+++ CHANGELOG.md
@@ -1,5 +1,21 @@
 # Release Notes
 
+## v1.0.5 (2024-02-23)
+
+- daisyui navbar + daisyui Sidebar Layout
+
+## v1.0.4 (2024-02-21)
+
+- daisyui navbar + Responsive Sidebar Layout native
+
+## v1.0.3 (2024-02-21)
+
+- daisyui
+
+## v1.0.2 (2024-02-21)
+
+- tailwindcss
+
 ## v1.0.1 (2024-02-21)
 
 - change from typescript to javascript